MB-8903: Get high sequence number as stat
Instead of getting it from the fake UPR server directly via
a function call, request it via a stat request. This currently
doesn't work with the real ep-engine implementation, as you
can't request a stat from a socket that is connected to an
UPR stream. You would need to open a separate socket.

As this still needs to discussed, the simplest possible way
(using the UPR stream socket also for stats) is used until
an agreement is found.

The 03-fake-upr-server.t tests are removed as all it's
functionality is testes with the 02-couch-upr.t tests anyway.

Change-Id: I0ae7bf74c71e2cc65702ce379458339b70f518c3