Make it possible to return several snapshots
It's now possible in the fake UPR server to return one than more
snapshot. The number of mutations within a snapshot can be specified.
In case the value is set >0 it's used. It will override the persisted
items function. This means that if the number if mutations per snapshot
is set, the returned on-disk snapshot will also only include that much
mutations.

Change-Id: I98de1d026d77bb9b2fef0fe733e1dc7fc91926ff