MB-11919: Limit updater starts
A `stale=update_after` starts the updater if it isn't already running.
When a lot of such requests are performed and the insertion rate is just
as high as the indexer can keep up with, then each run of the updater is
very short as it will need to process only a very small number of items.

The smaller the number of items the updater processes, the bigger the
fragmentation will be. Hence it's good to keep it at a reasonable size.
Especially during rebalance the updater processes extremely small batches.

This commit implements throttling for the updater on `stale=update_after`
requests. The updater will only be started if it wasn't started already
within a certain interval (currently 500ms). This will make the batches
during rebalance bigger and lead to less fragmentation.

Change-Id: I80b2c5c58f221d1eb4c53b54d72f60a6b07e7860