MB-13948 Queue items directly
Instead of building up a list of items, queue them directly into the
write queue. This reduces the memory usage significantly when there
are a lot of key-value emits per document.

The reason for this is that the map queue is bound to the size of the
documents and not to the size of the emits. This means that a lot of
emits can lead to not anticipated sizes.

It can be simulated with the following view map function:

    function (doc, meta) {
        for(var i=0; i<100; i++) {
            emit([meta.id, i], null);
        }
    }

It emits 100 key-values per document. For small document sizes, for
example with a value size of 10 bytes, this view function lead to a
memory usage of about 2GB on my machine. With this commit applied it
only took around 230MB.

Change-Id: Ib5102acbf4575794a5b6f5948c1dac3df5ded5e2