MB-18955 Compress temporary index files
For indexing the emitted key values are stored in temporary files
which will then be used by couchstore to insert them into the views.
Compressing those files saves a lot of disk space (up to 4x) and even
improves the overall indexing performance.

Several key values are combined before compression. Those groups are
then written as one compressed chunk to disk. It is prefixed by the
length of the chunk, followed by the compressed data itself.

Change-Id: I9ca7a800e9c79e11e23b216ef24d332eea43019e